DM1638 driver in System Verilog (1)
Extended the SPI command/data from 17 bits to 18 bits. The highest
bit indicates an additioonal instruction for the SPI driver. It tells
the driver a direction of the TM1638 command - writing the command followed
by an optional data byte, or writing the command and reading 4 bytes from
the device after that.

Migrated some type definitions from using 'reg' to use 'logic'.